# docs/swagger.yaml
basePath: /
definitions:
  error.errorResponse:
    properties:
      message:
        type: string
    type: object
  error.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  model.CommitPost:
    properties:
      post_commit:
        default: Wonderful
        type: string
      post_id:
        type: integer
    type: object
  model.Post:
    properties:
      tags:
        items:
          type: string
        type: array
      text:
        default: Hello World
        type: string
      title:
        default: Tutorial Golang
        type: string
    type: object
...
  model.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: https://mediumuz.herokuapp.com
info:
  contact:
    email: phapp0224mb@gmail.com
    name: Bakhodir Yashin Mansur
  description: API Server for MediumuZ Application
  termsOfService: https://github.com/BakhodiribnYashinibnMansur/Medium_uz
  title: MediumuZ API
  version: "1.0"
paths:
...
  /api/account/upload-image:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload Account Image
      operationId: upload-image
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload Account Image
      tags:
      - Account
...
  /api/ghost/post/get-commit:
    get:
      consumes:
      - application/json
      description: get commits
      operationId: get-commits
      parameters:
      - description: 'Offset '
        in: query
        name: offset
        type: integer
      - description: 'Limit '
        in: query
        name: limit
        type: integer
      - description: 'postID '
        in: query
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      summary: Get Commit
      tags:
      - Post
  /api/ghost/post/get/{id}:
    get:
      consumes:
      - application/json
      description: get post by id
      operationId: get-post-id
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      summary: Get  Post By ID
      tags:
      - Post
...
  /api/post/commit:
    post:
      consumes:
      - application/json
      description: Commit post by user
      operationId: commit-post-id
      parameters:
      - description: commit info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CommitPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Commit  Post
      tags:
      - Post
  /api/post/create:
    post:
      consumes:
      - application/json
      description: create post
      operationId: create-post
      parameters:
      - description: post info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Post
      tags:
      - Post
  /api/post/delete:
    delete:
      consumes:
      - application/json
      description: Delete post by id
      operationId: delete-post-id
      parameters:
      - description: Param ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete  Post By ID
      tags:
      - Post
  /api/post/like:
    get:
      consumes:
      - application/json
      description: Like post by id
      operationId: like-post-id
      parameters:
      - description: Param ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like  Post By ID
      tags:
      - Post
  /api/post/rating:
    get:
      consumes:
      - application/json
      description: Rated post by id
      operationId: rated-post-id
      parameters:
      - description: Param ID
        in: query
        name: id
        required: true
        type: integer
      - description: Rating Number
        in: query
        name: rating
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Rated  Post By ID
      tags:
      - Post
  /api/post/update:
    put:
      consumes:
      - application/json
      description: Update post by id
      operationId: update-post-id
      parameters:
      - description: Param ID
        in: query
        name: id
        required: true
        type: integer
      - description: post info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update  Post By ID
      tags:
      - Post
  /api/post/upload-image:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload Post Image
      operationId: upload-image-post
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload Post Image
      tags:
      - Post
  /api/post/view:
    get:
      consumes:
      - application/json
      description: View post by id
      operationId: view-post-id
      parameters:
      - description: Param ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: View  Post By ID
      tags:
      - Post
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"